refactor(tests): Move project tests to separate file
